𘥲	pldmm
This one is a bit messy, we have

𗋵      elqi
𗋵      eldqxi
𗌇      elqk
𗌇      elbrk
𗖑      svldq
𗖑      snvldq
𗤊      aldm
𗤊      aldmm
𗮪      culdm
𗮪      cbuldmm

where we used ldq, lb, ldmm to represent the same component 𘥲, although it is tempted to represent 𘥲 as lb, it is not very intuitive. Given that we have similar components 𘡶,𘣴, etc. that used d to represent the 𘠑 structure. I think `ldmm` is an acceptable representation.